Add first modules
