+++ .gitignore
@@ -1,2 +1,2 @@
 # .gitignore
-node_modules/
+node_modules
