+++ README.md
@@ -1,6 +1,6 @@
 ![](https://img.shields.io/badge/Microverse-blueviolet)
 
-# Project Name: Awesome Books
+# Project Name: Awesome Books using ES6
 
 > Description the project: With the design of this project the user can add books to the shelve (object array) by clicking the add button and remove the books when the remove button is clicked.
 
@@ -13,24 +13,11 @@
 
 ## Authors
 
-👤 **Ernest Anyewe Adonu**
-
-- GitHub: [@principlebrothers](https://github.com/principlebrothers)
-- Twitter: [@adonu_ernest](https://twitter.com/adonu_ernest)
-- LinkedIn: [Ernest Anyewe Adonu](www.linkedin.com/in/ernest-adonu-7b61951b0)
-
 👤 **Sajjad Rajaee**
 
 - GitHub: [@sajjadrajaee](https://github.com/sajjadrajaee)
 - Twitter: [@sajjadrajaee](https://twitter.com/sajjadrajaee)
 - LinkedIn: [LinkedIn](https://linkedin.com/in/sajjadrajaee)
-
-👤 **Seth Bizimana**
-
-- GitHub: [@Github](https://github.com/Sevenpros)
-- Twitter: [@twitter](https://twitter.com/BizimanaSeth)
-- LinkedIn: [LinkedIn](https://linkedin.com/in/sethBizimana)
-
 
 ## 🤝 Contributing
