+++ index.html
@@ -77,7 +77,7 @@
     </footer>
   </div>
 
-  <script src="index.js"></script>
+  <script type="module" src="index.js"></script>
 </body>
 
 </html>
