+++ modules/getdata.js
@@ -1,0 +1,9 @@
+function getData() {
+  const receivedData = localStorage.getItem('form');
+  if (receivedData) {
+    return JSON.parse(receivedData);
+  }
+  return JSON.parse(receivedData);
+}
+
+export default getData;
