+++ modules/storedata.js
@@ -1,0 +1,5 @@
+function storedData(book) {
+  localStorage.setItem('form', JSON.stringify(book));
+}
+
+export default storedData;
